+++ geotrek/core/apps.py
@@ -5,7 +5,7 @@ from django.utils.translation import gettext_lazy as _
 
 class CoreConfig(AppConfig):
     name = 'geotrek.core'
-    verbose_name = _("Core")
+    verbose_name = _("Topology")
 
     def ready(self):
         from .forms import PathForm, TrailForm