+++ docs/changelog.rst
@@ -1,5 +1,5 @@
 =========
-CHANGELOG
+Changelog
 =========
 
 2.102.1+dev (XXXX-XX-XX)