+++ docs/changelog.rst
@@ -9,13 +9,13 @@ CHANGELOG
 
 - Add `include_externals` filter to Cirkwi trek exports, to allow excluding treks with an external id (eid) (#3947)
 - Tourism : add price to TouristicEvent model - ref #3587
+- Add `check_versions` command to check Python, Django, PostgreSQL and PostGIS versions.
 
 **Bug fixes**
 
 - Fix cache key for zoning cities
 
-
-**Improvments**
+**Improvements**
 
 - Add popup button to add organizer in touristic event form
 - Change the `organizer` field of `TouristicEvent` model to a many to many field named `organizers` (#3587)