+++ docs/changelog.rst
@@ -10,6 +10,7 @@ CHANGELOG
 - Add `include_externals` filter to Cirkwi trek exports, to allow excluding treks with an external id (eid) (#3947)
 - Tourism : add price to TouristicEvent model - ref #3587
 - Add `check_versions` command to check Geotrek, Python, Django, PostgreSQL and PostGIS versions.
+- Add ``GeotrekCourseParser`` and ``GeotrekSiteParser`` in Aggregator to retrieve Outdoor models (refs #3569)
 
 **Bug fixes**